requirements.txt
    fastapi
    uvicorn[standard]
    mysql-connector-python 
    python-jose
    bcrypt==4.3.0